Add a VisualStudio Test Utility to prevent unnecessary recompilation
