+++ test/NuGet.Core.Tests/NuGet.PackageManagement.Test/PackageManagement/TestDependencyInfoProvider.cs
