+++ test/NuGet.Core.Tests/NuGet.PackageManagement.Test/PackageManagement/TestMetadataProvider.cs
