+++ test/NuGet.Core.Tests/NuGet.PackageManagement.Test/TestSourceControlManager.cs
