+++ test/NuGet.Core.Tests/NuGet.PackageManagement.Test/TestSourceControlManagerProvider.cs
