+++ test/TestUtilities/VisualStudio.Test.Utility/TestCpsPackageReferenceProject.cs
