+++ test/TestUtilities/VisualStudio.Test.Utility/TestExternalProjectReference.cs
