+++ test/TestUtilities/VisualStudio.Test.Utility/VisualStudio.Test.Utility.csproj
@@ -1,13 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>$(TargetFrameworksUnitTest)</TargetFrameworks>
-    <NoWarn>$(NoWarn);CS1591</NoWarn>
-    <NoWarn Condition=" $(TargetFramework.StartsWith('netcoreapp')) ">$(NoWarn);CS1998</NoWarn>
+    <TargetFrameworks>$(NETFXTargetFramework)</TargetFrameworks>
     <SkipShared>true</SkipShared>
     <TestProject>false</TestProject>
-    <SignWithTestKey>true</SignWithTestKey>
     <NoWarn>$(NoWarn);VSTHRD001;VSTHRD002;VSTHRD110</NoWarn>
-    <Description>A collection of test utilities, such as generating packages, mocking servers, stub implementations of interfaces, etc.</Description>
+    <Description>A collection of test utilities for Visual Studio NuGet extensions.</Description>
     <SkipAnalyzers>true</SkipAnalyzers>
     <UsePublicApiAnalyzer>false</UsePublicApiAnalyzer>
   </PropertyGroup>
@@ -17,33 +14,22 @@
     <Compile Include="$(SharedDirectory)\RequiredModifierAttributes.cs" />
   </ItemGroup>
 
-  <ItemGroup>
-    <InternalsVisibleTo Include="Dotnet.Integration.Test" />
-    <InternalsVisibleTo Include="NuGet.CommandLine.FuncTest" />
-    <InternalsVisibleTo Include="NuGet.Commands.FuncTest" />
-    <InternalsVisibleTo Include="NuGet.Commands.Test" />
-    <InternalsVisibleTo Include="NuGet.MSSigning.Extensions.FuncTest" />
-    <InternalsVisibleTo Include="NuGet.Packaging.FuncTest" />
-    <InternalsVisibleTo Include="NuGet.Packaging.Test" />
-    <InternalsVisibleTo Include="NuGet.Signing.CrossFramework.Test" />
+  <ItemGroup Condition=" '$(TargetFramework)' == '$(NETFXTargetFramework)' ">
+    <ProjectReference Include="..\..\..\src\NuGet.Clients\NuGet.PackageManagement.VisualStudio\NuGet.PackageManagement.VisualStudio.csproj" />
   </ItemGroup>
 
  <ItemGroup>
     <ProjectReference Include="..\..\..\src\NuGet.Core\NuGet.PackageManagement\NuGet.PackageManagement.csproj" />
     <ProjectReference Include="..\Microsoft.Internal.NuGet.Testing.SignedPackages\Microsoft.Internal.NuGet.Testing.SignedPackages.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Moq" />
-    <PackageReference Include="xunit" />
+    <ProjectReference Include="..\Test.Utility\Test.Utility.csproj" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == '$(NETFXTargetFramework)' ">
-    <PackageReference Include="AwesomeAssertions" />
     <PackageReference Include="Microsoft.Build" ExcludeAssets="runtime" />
     <PackageReference Include="Microsoft.Build.Tasks.Core" ExcludeAssets="runtime" />
     <PackageReference Include="Microsoft.Build.Utilities.Core" ExcludeAssets="runtime" />
-    <PackageReference Include="System.Formats.Asn1" />
+    <PackageReference Include="Microsoft.VisualStudio.ProjectSystem" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.VisualStudio.Sdk" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == '$(NETFXTargetFramework)' ">
@@ -52,9 +38,6 @@
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Net" />
     <Reference Include="System.Security" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' != '$(NETFXTargetFramework)' ">
-    <PackageReference Include="System.Security.Cryptography.Pkcs" />
+    <Reference Include="WindowsBase" />
   </ItemGroup>
 </Project>
